Updated Cog Files
Separated out existing cogs to different files for different types of commands as well as a dedicated testing cog as I kept breaking the other cogs when I was testing them. 

Added a command for sending Text-To-Speech messages and changed the formatting of various messages to show as embeds and pick a random color for the display.